requirements.txt
torch>=2.2
PyYAML>=6.0
numpy>=1.26
